Fix URL concatenation for Keycloak
This fixes an issue with how URLs are concatenated in OAuth2/Faraday

This lead to not being able to specify an URL prefix in the `site`
configuration attribute, ala `https://example.com/auth` which is
something that seems to be common on keycloak instances.

This commit specifically fixes the keycloak config by removing the
leading slash. I left everything else  as I am not sure if this would be
the correct thing to do for other providers.

(Specifically, the "custom" provider does not prefix URLs with a slash,
so this should be fine)